requirements.txt
langchain==0.1.17
faiss-cpu==1.7.4
google-generativeai==0.3.2
PyMuPDF==1.23.21
pdfminer.six==20221105
python-dotenv==1.0.1
